fix: paper arch related fixes
