+++ apps/paper/package.json
@@ -26,7 +26,7 @@
     "@babel/core": "^7.19.3",
     "@babel/runtime": "^7.19.4",
     "@react-native-community/eslint-config": "3.1.0",
-    "eslint": "^7.32.0",
+    "eslint": "^8.25.0",
     "metro-react-native-babel-preset": "^0.72.3"
   }
 }
