+++ packages/core/android/build.gradle
@@ -60,7 +60,7 @@ android {
 
     sourceSets {
         main {
-            if (isNewArchitectureEnabled() || project.properties['lottiereactnative_newArch'] == "true") {
+            if (isNewArchitectureEnabled()) {
                 java.srcDirs += ['src/newarch', "${project.buildDir}/generated/source/codegen/java"]
             } else {
                 java.srcDirs += ['src/oldarch']
