+++ packages/core/android/gradle.properties
@@ -25,4 +25,3 @@ android.useAndroidX=true
 android.enableJetifier=true
 
 lottiereactnative_kotlinVersion=1.7.10
-lottiereactnative_newArch=false
